+++ docs/release_notes.md
@@ -13,6 +13,7 @@
 ### ✨ New Functionality
 
 - [Orchestration] Added `GEMINI_3_1_FLASH_LITE` to model list in `OrchestrationAiModel`.
+- [Batch] Added a new `BatchesApi` client to access the [Batch Service API](https://github.tools.sap/AI/llm-batch-service).
 
 ### 📈 Improvements